requirements.txt
pydantic>=2.5
python-dotenv>=1.0
openai>=1.0
fastapi>=0.110
uvicorn[standard]>=0.27
